feat : service 완

+++ src/main/java/com/louter/collab/domain/page/dto/request/PageBlockEditRequest.java
@@ -8,5 +8,5 @@ public class PageBlockEditRequest {
 
     private String type;
 
-    private Long orderIndex;
+    private Integer orderIndex;
 }
